00_environment/05_project_structure/requirements.txt
mysqlclient==1.3.10
flask_sqlalchemy==2.2
flask_restplus==0.10.1
